requirements.txt
fastapi~=0.111.0
uvicorn~=0.29.0
tensorflow~=2.15.0
Pillow~=10.3.0
jinja2
numpy~=1.26.4